tests/wpt/tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.put.null.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.imageData.put.null</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.put.null</h1>
<p class="desc">putImageData() with null imagedata throws TypeError</p>


<script>
var t = async_test("putImageData() with null imagedata throws TypeError");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_js(TypeError, function() { ctx.putImageData(null, 0, 0); });
  t.done();

});
</script>
